+++ src/onegov.core/HISTORY.rst
@@ -3,6 +3,10 @@ Changelog
 
 Unreleased
 ~~~~~~~~~~
+
+- Adds a method which takes a list of dicts and turns it into a csv string.
+  [href]
+
 0.8.1 (2015-12-08)
 ~~~~~~~~~~~~~~~~~~~